Split settings into top level menu items
