+++ app/lib/features/settings/pages/chat_settings_page.dart
@@ -1,0 +1,44 @@
+import 'package:acter/common/extensions/acter_build_context.dart';
+import 'package:acter/common/widgets/with_sidebar.dart';
+import 'package:acter/features/settings/pages/settings_page.dart';
+import 'package:acter/features/settings/widgets/auto_download_tile.dart';
+import 'package:acter/features/settings/widgets/typing_notice_tile.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/l10n.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:settings_ui/settings_ui.dart';
+
+class ChatSettingsPage extends ConsumerWidget {
+  const ChatSettingsPage({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final lang = L10n.of(context);
+    return WithSidebar(
+      sidebar: const SettingsPage(),
+      child: Scaffold(
+        appBar: AppBar(
+          title: Text(lang.chat),
+          automaticallyImplyLeading: !context.isLargeScreen,
+        ),
+        body: SettingsList(
+          sections: [
+            SettingsSection(
+              tiles: [
+                CustomSettingsTile(child: AutoDownloadTile()),
+                CustomSettingsTile(child: TypingNoticeTile()),
+                SettingsTile.switchTile(
+                  title: Text(lang.chatSettingsReadReceipts),
+                  description: Text(lang.chatSettingsReadReceiptsExplainer),
+                  enabled: false,
+                  initialValue: false,
+                  onToggle: (newVal) {},
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
