+++ app/lib/router/shell_routers/home_shell_router.dart
@@ -20,8 +20,10 @@ import 'package:acter/features/profile/pages/my_profile_page.dart';
 import 'package:acter/features/public_room_search/pages/search_public_directory.dart';
 import 'package:acter/features/settings/pages/backup_page.dart';
 import 'package:acter/features/settings/pages/blocked_users.dart';
+import 'package:acter/features/settings/pages/calendar_settings_page.dart';
 import 'package:acter/features/settings/pages/change_password.dart';
-import 'package:acter/features/settings/pages/behavior_settings_page.dart';
+import 'package:acter/features/settings/pages/chat_settings_page.dart';
+import 'package:acter/features/settings/pages/customizations_settings_page.dart';
 import 'package:acter/features/settings/pages/email_addresses.dart';
 import 'package:acter/features/settings/pages/info_page.dart';
 import 'package:acter/features/settings/pages/labs_page.dart';
@@ -98,13 +100,35 @@ final homeShellRoutes = [
     },
   ),
   GoRoute(
-    name: Routes.settingsBehavior.name,
-    path: Routes.settingsBehavior.route,
+    name: Routes.settingsChat.name,
+    path: Routes.settingsChat.route,
     redirect: authGuardRedirect,
     pageBuilder: (context, state) {
       return MaterialPage(
         key: state.pageKey,
-        child: const BehaviorSettingsPage(),
+        child: const ChatSettingsPage(),
+      );
+    },
+  ),
+  GoRoute(
+    name: Routes.settingsCalendar.name,
+    path: Routes.settingsCalendar.route,
+    redirect: authGuardRedirect,
+    pageBuilder: (context, state) {
+      return MaterialPage(
+        key: state.pageKey,
+        child: const CalendarSettingsPage(),
+      );
+    },
+  ),
+  GoRoute(
+    name: Routes.settingsCustomizations.name,
+    path: Routes.settingsCustomizations.route,
+    redirect: authGuardRedirect,
+    pageBuilder: (context, state) {
+      return MaterialPage(
+        key: state.pageKey,
+        child: const CustomizationsSettingsPage(),
       );
     },
   ),
